Backend/Arman-Shaikh58_My-Portfolio/requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
